Update privacy.md
Removed: 'The government’s use of information about people is subject to a number of laws and policies, including: the Privacy Act of 1974, the Federal Information Security Management Act of 2002, and the eGovernment Act of 2002. Conduct a Privacy Threshold Analysis in collaboration your agency’s privacy office whenever a design calls for the creation of a new data store (for example, a database). Ensure all collections of personally identifiable information (PII) are accompanied by a Privacy Act Notice. See, for example, 18F’s Privacy Act Notice for Design Research.'

Removed: 'Familiarize yourself with the Fair Information Practice Principles, a set of precepts at the heart of the U.S. Privacy Act of 1974.'

Removed: 'which may include your general counsel'

Changed 'Government' to 'DAWE'. Updated with approved privacy and consent information.